Fix readRate/writeRate (long instead of int)
The readRate and writeRate stats from VDSM are gathered via libvirt.
Libvirt returns a long long (64bit long)
So we need to use Long in the engine to store this values
instead of int.

Bug-Url: https://bugzilla.redhat.com/2007384

+++ backend/manager/modules/common/src/main/java/org/ovirt/engine/core/common/businessentities/storage/DiskImage.java
@@ -22,9 +22,9 @@ public class DiskImage extends DiskImageBase {
     private long actualSizeInBytes;
     private long apparentSizeInBytes;
     private Long initialSizeInBytes;
-    private int readRateFromDiskImageDynamic;
+    private long readRateFromDiskImageDynamic;
     private long readOpsFromDiskImageDynamic;
-    private int writeRateFromDiskImageDynamic;
+    private long writeRateFromDiskImageDynamic;
     private long writeOpsFromDiskImageDynamic;
 
     // Latency fields from DiskImageDynamic which are measured in seconds.
@@ -226,11 +226,11 @@ public class DiskImage extends DiskImageBase {
         this.apparentSizeInBytes = apparentSizeInBytes;
     }
 
-    public int getReadRate() {
+    public long getReadRate() {
         return readRateFromDiskImageDynamic;
     }
 
-    public void setReadRate(int readRateFromDiskImageDynamic) {
+    public void setReadRate(long readRateFromDiskImageDynamic) {
         this.readRateFromDiskImageDynamic = readRateFromDiskImageDynamic;
     }
 
@@ -242,11 +242,11 @@ public class DiskImage extends DiskImageBase {
         this.readOpsFromDiskImageDynamic = readOpsFromDiskImageDynamic;
     }
 
-    public int getWriteRate() {
+    public long getWriteRate() {
         return writeRateFromDiskImageDynamic;
     }
 
-    public void setWriteRate(int writeRateFromDiskImageDynamic) {
+    public void setWriteRate(long writeRateFromDiskImageDynamic) {
         this.writeRateFromDiskImageDynamic = writeRateFromDiskImageDynamic;
     }
